+++ examples/docs-site/src/app/page.tsx
@@ -2,6 +2,7 @@ import { HeroSection } from '@/components/HeroSection'
 import { InstallSection } from '@/components/InstallSection'
 import { PatternSection } from '@/components/PatternSection'
 import { PrimitivesSection } from '@/components/PrimitivesSection'
+import { CustomizationSection } from '@/components/CustomizationSection'
 import { CliSection } from '@/components/CliSection'
 import { FooterSection } from '@/components/FooterSection'
 
@@ -12,6 +13,7 @@ export default function Home() {
       <InstallSection />
       <PatternSection />
       <PrimitivesSection />
+      <CustomizationSection />
       <CliSection />
       <FooterSection />
     </main>